requirements.txt
aiogram==3.21.0
python-dotenv==1.1.1
tweepy==4.16.0
ffmpeg-python==0.2.0